feat: Add authentication to routes

+++ src/routes/memories.ts
@@ -1,14 +1,31 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
 import z from 'zod'
+import { REPLServer } from 'repl'
 
 export async function routes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+
+  /*
+   funciona mais ou menos como um middleware
+   entao antes de qualquer requisicao, sera verificado se o usuario esta logado
+   ou se o token valido
+  */
+  app.addHook('preHandler',async (request) => {
+    await request.jwtVerify()
+  })
+
+
+  app.get('/memories', async (request) => {
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub
+      },
       orderBy: {
         createdAt: 'asc',
       },
     })
+
     return memories.map((memory) => {
       // retornando apenas algumas das linhas de texto, salvas na memoria.
       // do jeito que esta, é retornado um objeto mais leve, pois contém no maximo
@@ -21,7 +38,7 @@ export async function routes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
@@ -32,6 +49,10 @@ export async function routes(app: FastifyInstance) {
         id,
       },
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -49,14 +70,14 @@ export async function routes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'b8cf27b4-8fe9-41c4-9e93-8aea08f4b82b',
+        userId: request.user.sub,
       },
     })
 
     return reply.code(201).send(memory)
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
@@ -67,7 +88,18 @@ export async function routes(app: FastifyInstance) {
     })
     const { id } = paramSchema.parse(request.params)
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-    await prisma.memory.update({
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if(memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -75,14 +107,28 @@ export async function routes(app: FastifyInstance) {
         isPublic,
       },
     })
+
+    return memory
   })
 
-  app.delete('/memories/:id', async (request): Promise<void> => {
+  app.delete('/memories/:id', async (request, reply): Promise<void> => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramSchema.parse(request.params)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if(memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+
     await prisma.memory.delete({
       where: {
         id,
